Generate stuff for PQClean

+++ sha256-avx2/wots.c
@@ -138,7 +138,7 @@ static void base_w(unsigned int *output, const int out_len, const unsigned char 
             bits += 8;
         }
         bits -= SPX_WOTS_LOGW;
-        output[out] = (total >> bits) & (SPX_WOTS_W - 1);
+        output[out] = (unsigned int)(total >> bits) & (SPX_WOTS_W - 1);
         out++;
     }
 }
